+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ccl/dccwalletinfo/model/DccWalletInfoWrapper.kt
@@ -1,15 +1,15 @@
 package de.rki.coronawarnapp.ccl.dccwalletinfo.model
 
-import de.rki.coronawarnapp.ccl.dccwalletinfo.text.textResource
-import de.rki.coronawarnapp.ccl.dccwalletinfo.text.urlResource
+import de.rki.coronawarnapp.ccl.ui.text.urlResource
+import de.rki.coronawarnapp.ccl.ui.text.textResource
 
 data class DccWalletInfoWrapper(
     val dccWalletInfo: DccWalletInfo
 ) {
     val admissionState = dccWalletInfo.admissionState
-    val admissionBadgeText = ""// by textResource(admissionState.badgeText)
-    val admissionTitleText = ""// textResource(admissionState.titleText)
-    val admissionSubtitleText = ""// textResource(admissionState.subtitleText)
-    val admissionLongText = ""// textResource(admissionState.longText)
-    val admissionFaqAnchor = ""// urlResource(admissionState.faqAnchor)
+    val admissionBadgeText by textResource(admissionState.badgeText)
+    val admissionTitleText by textResource(admissionState.titleText)
+    val admissionSubtitleText by textResource(admissionState.subtitleText)
+    val admissionLongText by textResource(admissionState.longText)
+    val admissionFaqAnchor by urlResource(admissionState.faqAnchor)
 }
