+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/ccl/ui/text/UrlResourceKtTest.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.ccl.dccwalletinfo.text
+package de.rki.coronawarnapp.ccl.ui.text
 
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
